FOGL-4591:  separated source files for readings_catalogue
